int -> size_t in entity.c

+++ src/entity.h
@@ -1,8 +1,10 @@
 #pragma once
 
+#include <stddef.h>
+
 typedef struct EntityID {
 	int value;
 } EntityID;
 
 EntityID EntityID_new(void);
-int EntityID_count(void);
+size_t EntityID_count(void);
